Refactor main.js: torrent list controller
